Move util functions in main to cli_utils.py
